refactor adapter timings
