+++ src/adapt/gro-adapter-generic-build.ts
@@ -1,5 +1,3 @@
-import {Timings} from '@feltcoop/felt/util/time.js';
-import {print_timings} from '@feltcoop/felt/util/print.js';
 import {strip_trailing_slash} from '@feltcoop/felt/util/path.js';
 
 import type {Adapter} from './adapter.js';
@@ -25,22 +23,16 @@ export const create_adapter = ({
 		adapt: async ({config, fs, dev, log}) => {
 			await fs.remove(dir);
 
-			const timings = new Timings(); // TODO probably move to task context
-
 			const build_config = config.builds.find((b) => b.name === build_name);
 			if (!build_config) {
 				throw Error(`Unknown build config: ${build_name}`);
 			}
 
-			const timing_to_copy_dist = timings.start('copy builds to dist');
 			await copy_dist(fs, build_config, dev, dir, log);
-			timing_to_copy_dist();
 
 			if (host_target === 'github_pages') {
 				await ensure_nojekyll(fs, dir);
 			}
-
-			print_timings(timings, log);
 		},
 	};
 };
